fix: update ExecutionContext type across LVBS, optee, and Windows runner crates
Update remaining crates that still used PtRegs where ExecutionContext
is now required after pr18's EnterShim type change:
- litebox_platform_lvbs: run_thread, run_thread_ref, reenter_thread_ref,
  ThreadContext, run_thread_inner, call_shim all accept ExecutionContext;
  switch_to_user still receives &PtRegs via ctx.regs
- litebox_runner_lvbs: PtRegs::default() -> ExecutionContext::default()
- litebox_runner_optee_on_linux_userland: PtRegs -> ExecutionContext
- litebox_runner_linux_on_windows_userland: PtRegs -> ExecutionContext
- litebox_platform_linux_userland: fix 32-bit st_ino (u32 vs u64) and
  tv_nsec (subsec_nanos as c_long with cast_lossless suppression)

+++ litebox_platform_linux_userland/src/lib.rs
@@ -797,7 +797,7 @@ impl LinuxUserland {
             path,
             rdev: stat_buf.st_rdev,
             dev: stat_buf.st_dev,
-            ino: stat_buf.st_ino,
+            ino: stat_buf.st_ino as u64,
         })
     }
 
@@ -1065,7 +1065,8 @@ impl LinuxUserland {
         let ts = timeout.map(|t| libc::timespec {
             #[allow(clippy::cast_possible_wrap)]
             tv_sec: t.as_secs() as libc::time_t,
-            tv_nsec: libc::c_long::from(t.subsec_nanos()),
+            #[allow(clippy::cast_possible_wrap, clippy::cast_lossless)]
+            tv_nsec: t.subsec_nanos() as libc::c_long,
         });
         let _ = unsafe {
             libc::ppoll(
